refactor v2 validator reference tests
